Use a unique ID for each rich text table editor
Currently we’re giving the same id to every rich text table editor on a page. Because of the way Handsontable cell editors are constructed, there’s one editor per table, but when we have multiple tables per page they’re all creating table cell editors with the same id. This seems to be behind the weirdness we’re seeing with multi-table page editing.

As far as I can tell, the field id that gets passed to our `initAtomicTable` function doesn’t easy pass through Handsontable (whos constructor takes an element that we lookup by that id) to the cell editor. So my fix here is to crawl up the dom to where I can see we have an id that’s unique to this table and use that.

I don’t know if this is the right approach, but it demonstrates a fix.