Stack Contact Us table on mobile
The new Contact Us table introduced in PR 7533 doesn't display properly
on mobile. The old AtomicTableBlock on the Contact Us page had both
"Stacked on mobile" and "Full-width" options selected, but these
options were incorrectly not set on the new ContactUsTable. This commit
fixes that.

To test, visit http://localhost:8000/about-us/contact-us/ in a mobile
browser and compare with production.