Fix fragile table functional test
One of the table functional tests relies on the existence of a
hardcoded document in the Wagtail database named
"cfpb_interested-vendor-instructions_fy2020.pdf". If this document
does not exist, the test will fail.

This document doesn't exist in our production database, and so this
test current fails when run against production or production-like
systems.

This commit changes the test to be more flexible and test the
functionality of document selection within a table as opposed to the
existence of a specific document.