Restore Wagtail fallback database search
With the upgrade to Wagtail 3.0 we noticed that search in the Wagtail
admin is no longer working as expected; specifically, autocomplete and
substring searches don't seem to work the way they used to.

This seems related to Wagtail issue 7720, fixed by Wagtail PR 9900 in
version 4.2, at least for the PostgreSQL database backend.

As a workaround for now, we can use the Wagtail "fallback" database
search which uses a simpler substring-based search. If we ever want to
use Wagtail search for anything besides the Wagtail admin, we will have
to revisit this change.