Remove indexed page revision and search_pagerevisions
This change removes our `IndexedPageRevision` proxy model and our `search_pagerevisions` management commands.

These were added in #5871 to support a particular discovery effort, however, they have not been used since, negatively impact the page indexing time if we perform that via Wagtail, and search configuration brought in with them has since changed.

As we upgrade to Wagtail 4, we’d need to either modify or remove this. I’ve chosen to remove it.