Temporarily skip Bandit test for `requests.get()`
Bandit [1.7.5 adds a test for calls to `requests.get()` that do not include a `timeout` parameter. This PR adds that test to our skipped tests.

We have several of these, and we need to address them, but that can happen without our linting being blocked for all PRs at the moment.