Add custom table for consumer reporting companies
We currently configure our generic AtomicTableBlock to create tables on
Consumer Reporting Companies pages like this one:

https://www.consumerfinance.gov/consumer-tools/credit-reports-and-scores/consumer-reporting-companies/companies-list/uconfirm/

This PR adds a new module called the ConsumerReportingCompanyTable,
which simplifies the editing needed to create such a table. It's added
to the DocumentDetailPage page type which is used for these tables.

The block contains three rich text fields for contact information:
website, phone, and mailing address. Each of these is required.