Update Cypress block inventory tests for new report
Because wagtail-inventory also now uses select2 widgets, we have to be a little creative. Luckily [Cypress has a blog post on exactly how to test Select2 widgets](https://www.cypress.io/blog/2020/03/20/working-with-select-elements-and-select2-widgets-in-cypress/). django-autocomplete-light doesn’t assign field-specific ids to the widgets, so we have to go on what we know the page to look like.