Simplify block inventory functional tests
Since the block inventory moved to Wagtail reports and django-autocomplete-light with Select2, there’s not a lot of functionality unique to wagtail-inventory/cf.gov to test that isn’t dependent on Wagtail and django-autocomplete-light/Select2.

This simplifies this test to avoid just testing Select2/Wagtail reports — it verifies that the block inventory is present and has results, and that’s it.