TODO: Conditionally chose revision class in migration utils
We have data migrations that use `migrate_page_types_and_fields` that apply before the Wagtail 4 migration from `PageRevision` to `Revision`, so we need to ensure we continue support for `PageRevision` until those migrations are squashed away.

This commit is marked as TODO because we should squash before merging Wagtail 4 support into main, and when we do we can edit this commit to just use `Revision`.