Fix language activation in page previews
Wagtail 4.0 decoupled the Page.serve and Page.serve_preview methods,
which means that serve() code that would previously run at preview time
no longer does:

https://docs.wagtail.org/en/stable/releases/4.0.html#changes-to-page-serve-and-page-serve-preview-methods

Practically, we override Page.serve in order to activate translations
for the page language. In Wagtail 4.0, we need to make sure that this
same code runs in Page.serve_preview. This commit does so.

To observe how this is broken on the current Wagtail 4 branch, edit
a non-English page (for example, on a production dump,
http://localhost:8000/admin/pages/17009/edit/) and look at the sidebar
preview. If you click in the site search in the header, you'll see the
blue search button says "Search" in English, which is wrong -- it
should be translated in the page's language.

With this commit, that "Search" term will be properly translated.