Change Django X-Frame-Options to SAMEORIGIN
Django currently sets the HTTP X-Frame-Options header to DENY. This
conflicts with a separate Apache configuration that sets this to
SAMEORIGIN.

The Django 3.2 baseline requires that this be set either to DENY or
SAMEORIGIN (see internal platform#4418 for reference).

Setting this to DENY prevents any Django-served pages from being
loaded in an iframe, which prevents the Wagtail 4 live preview panel
from working. To enable the live preview panel, this commit changes the
setting from DENY to SAMEORIGIN, which is still in compliance with the
Django baseline.