Deprecate custom Django admin logout template
When users logout from the Django admin (not to be confused with the
totally separate action of logging out from the Wagtail admin) they see
a custom template which is currently visually broken.

This commit removes that custom template and sets the Django setting
LOGOUT_REDIRECT_URL so that users log out and are taken back to the
Wagtail login view.