Apply latest migrations to test database
This PR applies the most recent set of migrations to our test database,
test.sql.gz. All I've done is a refresh-data.sh (which applies the
migrations) followed by a dump-data.sh.

The last time we updated the test database was on January 20th; since
then we've upgraded to Wagtail 3 and have a migration squash and
another Wagtail upgrade coming soon.