Hide Django admin link from non-staff users
Currently users who don't have permission to access the Django admin
(i.e. non-staff users) still see a "Django Admin" link in the Wagtail
sidebar, which points to /django-admin/.

If those users click on that link, they get stuck in an infinite
redirect loop: /django-admin/ bumps them to the Wagtail login with a
?next parameter, Wagtail sees that they're already logged in, and sends
them back the Django admin, etc.

It doesn't seem easy/possible to avoid that infinite loop unfortunately.
The best we can do is hide the Django admin link for those users, which
is what this PR does.