Update get_latest_revision_as_page calls
Wagtail 4.0 renamed the `get_latest_revision_as_page` method to
`get_latest_revision_as_object`; the old name still exists for
backwards compatibility, but will be removed in Wagtail 5:

https://docs.wagtail.org/en/stable/releases/4.0.html#page-get-latest-revision-as-page-renamed-to-page-get-latest-revision-as-object

We have a couple of references to the old name, which generate a
warning when running tests. This commit renames them.