Fix our env label
This change updates our env label for Wagtail’s new sidebar. Instead of sitting on top of the sidebar, it now sits beside the collapse button, and will move with the sidebar to collapse.

It still appears across the top on mobile, but now it appears dimensionally the same as the sidebar button.