Fix translation links when previewing new pages
Translation links are currently broken when you use the Wagtail 4+
live preview panel to preview new pages that have never been saved.
A translation link is shown for almost every page on the site. This is
because the current logic doesn't properly handle the case where a
page's primary key is null.

This commit fixes that logic, while still ensuring that the links
render properly both for new English and non-English pages.