Allow disabling StreamField block unescaping
By default we always unescape the rendering of all StreamField blocks;
this allows potentially unsafe HTML characters to be added to just
about any block, which is not great. See internal platform 3839.

This adds the ability to disable unescaping for particular blocks that
we know should never contain any raw HTML characters, and for which we
want to explicitly prohibit this from happening.

This can be done by setting an "unescape = False" property to the
block's Meta class, for example:

```
class MyBlock(blocks.StructBlock):
   ...

   class Meta:
      ...
      unescape = False
```

If this is not set, by default the current behavior will continue,
meaning that the block's rendered HTML will be unescaped before being
added to the page.