Deprecate unused fields from AbstractFilterPage
This commit removes two unused fields from AbstractFilterPage:
comments_close_by and date_filed. See this gist for some data
investigation and explanation about how these fields are unused:

https://gist.github.com/chosak/a1b3c4ff5707fc9f45454ce06abb2ad3

Additionally, this commit removes a currently-empty "Scheduled
Publishing" item from the page configuration tab; as of Wagtail 4 this
is now in the page editor sidebar and so we don't need to include it
explicitly as a section in our page settings.