Fix "first page date" logic
Commit 2d9839bfb1460c630a0d1de21bbe3e2f7b72b984 introduced a bug in how
we determine the default "first page date" when rendering filterable
list pages. Previously, we could assume that Elasticsearch results were
always sorted from earliest to latest, and so the code used
`all_filterable_results[0]` to get the earliest post. Now, we can't
assume that, so we have to find the earliest post ourselves.

To test, visit a URL like this:

http://localhost:8000/about-us/blog/?title=&from_date=&to_date=2021-01-15

If you don't specify a `from_date`, the code fills it in. On production,
this is currently wrong:

https://www.consumerfinance.gov/about-us/blog/?title=&from_date=&to_date=2021-01-15

but this change fixes that.