Cleanup unnecessary filterable list logic
Commit 19e69581d2614fe6a2c5dff1485a6d1564c11725 introduced a new
search aggregation for filterable lists to optimize computation of the
earliest post date in search results. Because of the way we use Django
caching as part of filterable search, this logic had to be deployed
incrementally such that existing cache values would work even without
having the aggregation value.

This code has been deployed and the cache has been cleared in all
environments; thus, the fallback logic can now be removed. This change
does so.