Remove 50/50 layout option in filterable lists
With the March 2022 retirement of

https://www.consumerfinance.gov/consumer-tools/everyone-has-a-story/

we no longer have any pages on the site that use the 50/50 layout
option when rendering filterable list results.

See snapshot version from archive.org here:

https://web.archive.org/web/20211227063058/https://www.consumerfinance.gov/consumer-tools/everyone-has-a-story/

This commit deprecates the `output_5050` option to render results that
way. This also allows for removal of the `link_image_and_heading`
option that was only ever used in the 50/50 configuration.