Deprecate unused get_latest_activities method
This commit removes an unused `get_latest_activities` method along with
its tests. This was added back in 2016 in commit
ad0b9f0586707c582afd9c9c3eaba8c3b77a21cd.

It looks like I removed the last use of this in 2018 in commit
9656d21ce3b64941acaae5cdb00818e5061dfd57
but neglected to remove the actual method itself.