Default PortalSearchPage results to alphabetical
Currently the results on PortalSearchPages like
https://www.consumerfinance.gov/consumer-tools/auto-loans/answers/basics/
default to a random order if the user hasn't entered a search query.

If they do enter a search query, the results are sorted by search
relevance (the default behavior of OpenSearch, using the `_score`
field).

This commit changes the default sort order to be alphabetical.