Upgrade OpenSearch-related libraries
This commit bumps the version of the opensearch-py package from 1.1.0
to 2.2.0. This allows us to remove our dependence on opensearch-dsl,
which was incorporated into opensearch-py in the latest version.

It also bumps the version of the django-opensearch-dsl from 0.4.1 to
0.5.1.

All unit and functional tests continue to pass with these upgrades.