Add keyword field index to PortalSearchPages
We currently index PortalSearchPages into OpenSearch so they can be
searched on pages like

https://www.consumerfinance.gov/consumer-tools/auto-loans/answers/basics/

The "autocomplete" field in the search index actually represents the
page title. This is currently indexed as a text field so that fuzzy
searches work against it.

But we also want to index this same title as a keyword field so that
we can sort on the field, i.e. so that search results can be shown in
alphabetical order.

This commit makes a change to the way this field is indexed so that it
gets indexed both as a text field AND as a keyword field (under the
name "raw").

A subsequent change will actually use this new field to add
alphabetical ordering; this needs to be done in two steps so that we
first ensure that all live environments get the new field before we
try to use it.

There's no real testing to be done besides trying this PR on an
existing DB and search index and visiting pages like:

http://localhost:8000/consumer-tools/auto-loans/answers/basics/
http://localhost:8000/consumer-tools/auto-loans/answers/basics/?search_term=credit

You won't see any changes to search results. Reindexing with this change
and then revisiting those URLs also won't show any differences.