Minor correction to EmailSignUp help text
Currently when editing an EmailSignUp snippet, users are supposed to
either fill in the GovDelivery code or the signup URL.

The code help text says: "GovDelivery Code (USCFPB_###) for the list
people who submit the form will sign up for. Provide either this or the
signup URL, but not both."

And the URL help text says: "URL for the GovDelivery signup page people
will be linked to in order to signup. Provide either this or the signup
URL, but not both."

But this second text is wrong - it should say "this or the GovDelivery
code, but not both."

This commit makes that correction.