Prototype preview fields audit tool
This commit adds a prototype tool for auditing a potential change to
page post preview fields, as discussed on internal D&CP#234.

It makes the following changes to pages and allows for comparison of
the post preview view:

- Remove `preview_title` in favor of `seo_title`
- Remove `preview_subheading` entirely
- Remove `preview_description` in favor of `search_description`
- Remove `preview_image` entirely, falling back on default CFGOVPage
behavior.

To test, run locally and visit:

http://localhost:8000/admin/reports/page-preview-comparison/

Pages with differences will be highlighed in red; note there may be
false alarms due to noise in the way that rich text data is stored.

This is just a first cut at a prototype tool and is far from polished.