Move some FilterableList fields to page classes
This commit copies some of the configuration settings from the
FilterableList StreamField block to the FilterableListMixin page class
mixin. It alters the FilterableListMixin logic to use these fields
instead of the values set on the FilterableList block. This commit also
includes a data migration to populate these values.

The configuration settings will be removed from the FL block in a
subsequent commit.