Correct references to squashed migrations
In previous attempts to squash our migrations, we neglected to remove
a few references to the old migration names. Per the Django docs [0],
we need to

> Updating all migrations that depend on the deleted migrations to
> depend on the squashed migration instead.

but it seems like we missed a few of these.

A way to check this is to try running something like

```
cfgov/manage.py sqlmigrate v1 0001
```

which is supposed to dump out the SQL for a migration; this won't work
on main because of these bad references.

[0] https://docs.djangoproject.com/en/4.2/topics/migrations/#squashing-migrations