Add necessary dependency to migration
The v1 0242 migration does a StreamField data migration and depends on
some Wagtail 4.x changes (specifically, that revision data is stored in
a Revision model as JSON), but the migration didn't include a
dependency on the relevant Wagtail migration. This could cause problems
in certain scenarios, for example when running initial-data.sh to
populate an empty database.