Automatically assign IDs to rendered headings
This change alters the template for the existing HeadingBlock so that
the heading tag (e.g. <h2>) always gets an `id` attribute with a
slugified version of the heading text.

This uses our existing `slugify_unique` method that avoids duplicate
IDs by enforcing unique IDs for a single Django request. For example,
if there are two headings with the same text ("Example"), the first
will will be given the ID "example" and the second "example-2".