Deprecate atomic.html table template
As part of the deprecation of AtomicTableBlock in commit
8f4a52850dfd875e4b27b831b5591cbcc0701485, most of our tables were moved
off of the atomic.html template to a new base table template. There
was one exception, the ContactUsTable block which is only used on one
page of the site:

https://www.consumerfinance.gov/about-us/contact-us/

As discussed internally on D&CP#310, we wanted to deprecate this legacy
atomic.html template so that all tables use the same base.html file.
This commit does so.

To test, see these template debug views:

http://localhost:8000/admin/template_debug/v1/table/
http://localhost:8000/admin/template_debug/v1/contact_us_table/

And also compare the production page against this local one:

http://localhost:8000/about-us/contact-us/