Deprecate reliance on Wagtail's testapp
Currently we use Wagtail's wagtail.test.testapp as part of our test
settings in order to use some of its test models, for example
SimplePage.

There's no need to use Wagtail's test models if we have a model that
can work just as well. Additionally, there's an issue with the
migrations for the testapp in Wagtail 4.2.4, and this avoids needing
to deal with that -- see wagtail#10724.

Theoretically we could maintain our own test models, for example using
the method documented here [0], but for now using our own models
works just as well and avoids the dependency on Wagtail's testapp.

[0] https://docs.djangoproject.com/en/4.2/topics/testing/tools/#isolating-apps