Clarify comment about fallback database backend
We have a TODO comment in the code about something that can be removed
in Wagtail 4.2, but it turns out that we can't actually remove this
setting. If we want to maintain the easy (meaning: we don't need to
worry about running a search index update job) substring search matching
behavior in Wagtail admin search, we need to keep this setting even on
Wagtail 4.2.

Instead of removing, I've instead clarified the comment.