Add "live" field to filterable search index
Currently the search index used for filterable pages does not include
the `live` field which indicates whether a page is live or not. This
means that search results could potentially include pages that aren't
currently live.

We encountered this on internal issue D&CP#324, where unpublished pages
appear in search results.

This commit adds the `live` field to the search index; a subsequent
commit will actually use the field to filter pages once we have ensured
that all environment indexes have been rebuilt.