Fix: filtering should only return live pages
This commit modifies the page filtering logic to ensure that only live
pages are returned from search. We have various other logic to try to
prevent not-live pages from being populated in the index, but there are
some edge cases where that might happen. This commit filters out any
not-live pages from the search.

This requires a re-indexing of the search index after commit
1a843280dfd3936e4f6d879196617ed662743b07.

Addresses internal D&CP#324.