Customize JobListingPage admin display title
This commit modifies the Wagtail admin display title of JobListingPages.

If a page has no grades assigned, its display title is unchanged.
With one grade, the title will appear like "Title [Grade 53]".
With multiple, it will appear like "Title [Grades 53, 60]".

Note that because of the way that Wagtail drafts work, this display
title will unfortunately not always show the grades as of the latest
Wagtail draft. It will show the _title_ as of the latest Wagtail draft,
but the _grades_ portion will be fixed to the latest published version,
or the first draft if the page has never been published.

Addresses internal D&CP#333.