TDP: No results view should match results view
To address internal D&CP#347, this change modifies the TDP activity
search view so that it has the same layout whether there are results
or not. Specifically, if there are no results, this change keeps the
facet sidebar and moves the "no results" text where the results would
otherwise go.

To test locally, rebuild the frontend with `./frontend.sh`, make sure
you have an up-to-date database and search index, and visit

http://localhost:8000/consumer-tools/educator-tools/youth-financial-education/teach/activities/?grade_level=5&activity_duration=3&topic=8

Try the page on desktop and mobile; when on mobile, tab through the
page elements and confirm that the "Skip to search results" jump link
does not appear when there are no results even though it does when
there are results.