Add report_raw_html management command
This change adds a management command designed to report possible raw HTML that has been entered into any `StreamField`s on a given page model.

It does so by regular expression pattern-matching a `&lt;` followed by any text that isn't `&gt;` that is then followed by `&gt;`. It uses a combination of Django/Postgres regular expression query filtering to filter out pages that match that regular expression in the database, and then after fetching them from the database, walks through the given field's JSON and returns a path to the field.

For JSON like:

```json
[{
    "type": "full_width_text",
    "value": [{
        "id": ""some hash",
        "type": "content",
        "value": "<p>Some rich text&lt;br&gt;here.</p>"
    }]
}]
```

The "path" will be:

```
0.full_width_text.0.content
```

This way we should be able to identify both

1. Where on a page (including possibly multiple different fields) raw HTML is used
2. What types of fields use raw HTML

This way we can eliminate the raw HTML from the page, *and* remove support for rendering it at the field level once all instances of specific field types are fixed.

Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>