Delete unused modeladmin template
This commit deletes an unused Wagtail template override
(modeladmin/create.html). This template was added back in 2017 as part
of the initial Ask CFPB implementation, which used Wagtail modeladmin
to manage Django Answer models. It was added in commit
31d62bc5a997300b40f4f736c48fced2d3f34057. This custom template was
intended to allow easier switching between English and Spanish when
creating new Answers.

A later reimplementation of Ask CFPB deprecated the Answer model in
favor of Wagtail AnswerPages. This also deprecated the use of modeladmin
in commit 7e53a93267c1e0882af0ba760fbaf101a5283061, but left this
template behind. It isn't being used anywhere and so is removed here.