Wrap expandable links in paragraph tag
This change alters the rendering of the Expandable block so that link
tags are rendered inside of a `<p>..</p>` element. This prevents
elements from incorrectly rendering directly next to each other.

See internal D&CP#392 for context.

To test, visit this page with a production dump to compare spacing
around link elements with the production version:

http://localhost:8000/consumer-tools/educator-tools/youth-financial-education/resources-research/building-block-activities-resources/

Additionally, try adding two Links inside an Expandable to see that
they are properly spaced.