Pin Pillow to 10.1.0 to make Snyk happy
Snyk reports a vulnerability in Pillow versions < 10.0.0. We don't
pin Pillow, which means that Wagtail in practice pulls in the most
recent version, which is greater than 10.0.0, BUT for some reason this
doesn't satisfy Snyk.

Snyk keeps reporting this vulnerability on all of our cf.gov PRs. In
order to make Snyk happy, let's pin Pillow to the currently installed
version, which is 10.1.0.