Convert DB unique index to unique constraint
This is a one-off migration used to convert an existing DB unique index
on a Wagtail core table into a unique constraint. This is needed to
enable upgrade to Wagtail 5.1.

Once this migration has been applied to all permanent cf.gov
environments, we can squash migrations and remove it.

See internal Design-and-Content-Team#389 for more context.