Stop hiding "Delete" option for pages
Wagtail 5.2 unifies some aspects of the admin UI such that page UI
elements are reused elsewhere, for example snippets. This makes it
harder to hide "Delete" for pages but not in other places where we
want it to appear.

We already prohibit deletion on the backend:

https://github.com/cfpb/consumerfinance.gov/blob/f6622906a8ef4825b1e82fb2f1df0b6abc3f0a5f/cfgov/v1/wagtail_hooks.py#L72-L98

and so this change doesn't actually make it possible for users to delete
pages. If they try to delete, they'll see an error: "Sorry, you do not
have permission to access this area."