Fix use of SKIP_DJANGO_MIGRATIONS without a DB
Running `SKIP_DJANGO_MIGRATIONS=1 tox -e unittest` runs the Python unit tests without running migrations, as documented here:

https://cfpb.github.io/consumerfinance.gov/python-unit-tests/#running-tests

If you try this to run a subset of tests that don't require a database (for example because they inherit from Django SimpleTestCase), you'll get a Django IntegrityError.

This commit fixes that by confirming that we're using a test database before trying to populate it with test data.