Add a regex-based alter value migration operation
This change adds a `RegexAlterBlockValueOperation`, based on Wagtail's `BaseBlockOperation` to perform a regular expression-based substitution on the value of a block in a `StreamField`. This is using Wagtail's new StreamField migration utilities.

The usage is to provide the operation in a `MigrateStreamData` operation in a data migration:

```python
class Migration(migrations.Migration):

    dependencies = [
        ("...", "...")
    ]

    operations = [
        MigrateStreamData(
            app_name="v1",
            model_name="LearnPage",
            field_name="content",
            operations_and_block_paths=[(
                RegexAlterBlockValueOperation(
                    search_pattern,
                    replacement_pattern
                ),
                "full_width_text.content"
            ),
        ]),
    ]
```