Corrections to late fee sorting logic
When sorting by late fees, we don't want to show cards that have late
fees but haven't specified a first late fee in dollars. Additionally,
if a card has no late fees but has a null late fee in dollars, we want
to display it as zero.

See internal DTUR#221 and internal DTUR#226 for background.