Isolate remaining raw HTML in UnescapedRichTextBlock
This change adds an `UnescapedRichTextBlock` subclass of `RichTextBlock` that explicitly converts HTML entities stored in the block value to HTML. This has the effect of supporting manually-entered "raw HTML". The block also validates against an allowed list of possible raw HTML tags, currently only permitting SVG `svg` and `path` elements.

This is the same process that happens in our `render_stream_child` function at present, but that will be removed in a later commit.